% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_transformation.R
\name{inv_log2_transform}
\alias{inv_log2_transform}
\title{Inverse of the scaled log2-transform of antibody titres}
\usage{
inv_log2_transform(log2_titre)
}
\arguments{
\item{log2_titre}{Numeric vector of log2-transformed titres.}
}
\value{
Numeric vector of original antibody titres.
}
\description{
Converts log2-transformed titres back to the original scale.
This is the inverse of `log2_transform()`. For example, a log2-transformed
titre of 1 corresponds to an original titre of 10, and each increment of 1
doubles the original titre.
}
\examples{
inv_log2_transform(1)  # 10
inv_log2_transform(2)  # 20
}
