% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infections_since_birth.R
\name{simulate_DENV_infections_via_competition}
\alias{simulate_DENV_infections_via_competition}
\title{Title}
\usage{
simulate_DENV_infections_via_competition(lambda, min_gap, stop_time, rho)
}
\arguments{
\item{lambda}{TBD}

\item{min_gap}{TBD}

\item{stop_time}{TBD}

\item{rho}{TBD}
}
\value{
a data frame
}
\description{
Title
}
\examples{
simulate_DENV_infections_via_competition(0.1, 365, 5 * 365, 0.01)
}
