% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_process.R
\name{simulate_infection_times}
\alias{simulate_infection_times}
\title{Simulate infection times}
\usage{
simulate_infection_times(daily_lambda, stop_time, min_gap, max_inf)
}
\arguments{
\item{daily_lambda}{A scalar corresponding to the daily force of infection.}

\item{stop_time}{A numeric scalar specifying the maximum time horizon for the
simulation.}

\item{min_gap}{A scalar representing the minimum time that must elapse
between two consecutive infections. This interval is due to temporary
immune protection.}

\item{max_inf}{An integer indicating the maximum number of infections that
an individual can experience from a given pathogen.}
}
\value{
A numeric vector of infection times occurring before \code{stop_time}
}
\description{
Simulate infection times
}
\examples{
simulate_infection_times(lambda     = 0.3,
                         stop_time  = 365 * 3,
                         min_gap    = 365,
                         max_inf    = 4)
}
