% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_process.R
\name{simulate_infection_times_since_birth}
\alias{simulate_infection_times_since_birth}
\title{Simulates infection times from birth}
\usage{
simulate_infection_times_since_birth(
  daily_lambda,
  age_enrolment,
  offset_start,
  stop_time,
  min_gap,
  max_inf
)
}
\arguments{
\item{daily_lambda}{A scalar corresponding to the daily force of infection.}

\item{age_enrolment}{Numeric. Age of the individual at study enrolment (in years).}

\item{offset_start}{Numeric. Time from start of study to enrolment (in days).}

\item{stop_time}{Numeric. Duration from the start to the end of the study (in days).}

\item{min_gap}{A scalar representing the minimum time that must elapse
between two consecutive infections. This interval is due to temporary
immune protection.}

\item{max_inf}{An integer indicating the maximum number of infections that
an individual can experience from a given pathogen.}
}
\value{
Numeric vector of infection times (relative to birth).
}
\description{
Simulates infection times from birth
}
\examples{
simulate_infection_times_since_birth(
lambda        = 0.3 / 365,
age_enrolment = 8, # in years
offset_start  = 0,
stop_time     = 365 * 5,
min_gap       = 365,
max_inf       = 4)

}
