% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_infections.R
\name{simulate_infection_times_since_susceptibility}
\alias{simulate_infection_times_since_susceptibility}
\title{Simulate infection times since susceptibility}
\usage{
simulate_infection_times_since_susceptibility(
  lambda,
  age_enrolment,
  enrolment_time,
  follow_up_time,
  min_gap,
  max_inf
)
}
\arguments{
\item{lambda}{A scalar corresponding to the annual force of infection.}

\item{age_enrolment}{Numeric. Age of the individual at study enrolment, in years.}

\item{enrolment_time}{Numeric. Time from the start of the study to this
individual's enrolment, in days.}

\item{follow_up_time}{Numeric. Time from the start of the study to the
individual's last measurement, in days.}

\item{min_gap}{A scalar representing the minimum time that must elapse
between two consecutive infections. This interval is due to temporary
immune protection.}

\item{max_inf}{An integer indicating the maximum number of infections that
an individual can experience from a given pathogen.}
}
\value{
Numeric vector of infection times, in days since becoming susceptible
  (that is, since the end of maternal antibody protection).
}
\description{
Simulates infection times from the end of maternal antibody protection
(assumed to occur 1 year after birth) until the individual's last
measurement.
}
\examples{
simulate_infection_times_since_susceptibility(
  lambda = 0.3,
  age_enrolment = 8,      # years
  enrolment_time = 0,     # days since study start
  follow_up_time = 365 * 5,
  min_gap = 365,
  max_inf = 4
)

}
