% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_trajectories.R
\name{simulate_titre_trajectory}
\alias{simulate_titre_trajectory}
\title{Simulate Antibody Titre Trajectories for an Individual}
\usage{
simulate_titre_trajectory(
  sampling_times,
  treatment_group,
  subject_id,
  meas_sd,
  infection_times = NULL,
  vac_times = NULL
)
}
\arguments{
\item{sampling_times}{Numeric vector of sampling times (in days) relative to
the start of the study.}

\item{treatment_group}{Integer (0 or 1) indicating treatment group.
(0 = Placebo, 1 = Vaccine).}

\item{subject_id}{Identifier for the subject (numeric or character).}

\item{meas_sd}{A non-negative real value that corresponds to the standard
deviation of a normal measurement model of log titres.}

\item{infection_times}{Numeric vector of times of infection events relative
to enrolment. Defaults to `NULL` if no infections.}

\item{vac_times}{Numeric vector of times of vaccination events relative to
enrolment. Defaults to `NULL` if no vaccinations.}
}
\value{
A `data.frame` with columns:
\describe{
  \item{subject_id}{Subject identifier.}
  \item{time}{Sampling time.}
  \item{true}{Simulated true log-titre at each time.}
  \item{meas}{Observed log-titre with measurement error and LOD applied.}
}
}
\description{
Generates a simulated time course of true and measured antibody titres
for an individual, given sampling times, infection/vaccination events,
and baseline serostatus. The function applies a rise-and-decay model
with measurement error and limit of detection.
}
\examples{
  sampling_times  <- c(182, 210, 266, 294, 434, 643, 980, 1347, 1740)
simulate_titre_trajectory(sampling_times   = sampling_times,
                         treatment_group  = 0, # Placebo
                         subject_id       = 1,
                         infection_times  = c(300, 900),
                         meas_sd          = 0.3)
}
