% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_trajectories.R
\name{simulate_titres_seropositive}
\alias{simulate_titres_seropositive}
\title{Simulate antibody titre trajectories from birth}
\usage{
simulate_titres_seropositive(
  inf_times_sbs,
  sampling_times,
  age,
  subject_id,
  peaks,
  perm_rises,
  decays,
  meas_sd
)
}
\arguments{
\item{inf_times_sbs}{Numeric vector of infection times (in days) relative to
day the individual became susceptible.}

\item{sampling_times}{Numeric vector of sampling times, in days relative to
the start of the study.}

\item{age}{Numeric. Age of the individual at the time of the first sample
(in years).}

\item{subject_id}{Subject identifier.}

\item{peaks}{Numeric vector of peak titres immediately after each exposure.
Must have the same length as `exposure_times`.}

\item{perm_rises}{Numeric vector of long-term plateau titre values following
each exposure. Must have the same length as `exposure_times`.}

\item{decays}{Numeric vector of decay rates. Must have the same length as
`exposure_times`.}

\item{meas_sd}{Non-negative numeric scalar. Standard deviation of the normal
measurement model for log titres.}
}
\value{
A data frame containing simulated titre values across the specified
 sampling times.
}
\description{
Simulate antibody titre trajectories from birth
}
\examples{
simulate_titres_seropositive(
  inf_times_sbs   = c(1059.646, 2743.543, 4152.091),
  sampling_times  = c(41, 74, 122, 157, 290, 468, 787, 1192, 1553),
  age             = 10,
  subject_id      = 3,
  treatment_group = 0,
  perm_rise       = 6,
  temp_rise       = 2,
  meas_sd         = 0)
}
