% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_trajectories.R
\name{simulate_titres_seropositive}
\alias{simulate_titres_seropositive}
\title{Simulate antibody titre trajectories from birth}
\usage{
simulate_titres_seropositive(
  inf_times_sbs,
  sampling_times,
  age,
  subject_id,
  treatment_group
)
}
\arguments{
\item{inf_times_sbs}{Numeric vector of infection times (in days) relative to
day the individual became susceptible.}

\item{age}{Numeric. Age of the individual at the time of the first sample
(in years).}
}
\value{
A data frame containing simulated titre values across the specified
 sampling times.
}
\description{
#' @inheritParams simulate_titre_trajectory
}
\examples{
simulate_titres_seropositive(
  inf_times_sbs   = c(1059.646, 2743.543, 4152.091),
  sampling_times  = c(41, 74, 122, 157, 290, 468, 787, 1192, 1553),
  age             = 10,
  subject_id      = 3,
  treatment_group = 0)
}
