% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_trajectories.R
\name{simulate_titres_seropositive}
\alias{simulate_titres_seropositive}
\title{Simulate antibody titre trajectories from birth}
\usage{
simulate_titres_seropositive(
  inf_times_sbs,
  sampling_times,
  age,
  subject_id,
  treatment_group,
  perm_rise,
  temp_rise,
  temp_decay,
  meas_sd
)
}
\arguments{
\item{inf_times_sbs}{Numeric vector of infection times (in days) relative to
day the individual became susceptible.}

\item{sampling_times}{Numeric vector of sampling times (in days) relative to
the start of the study.}

\item{age}{Numeric. Age of the individual at the time of the first sample
(in years).}

\item{subject_id}{Identifier for the subject (numeric or character).}

\item{treatment_group}{Integer (0 or 1) indicating treatment group.
(0 = Placebo, 1 = Vaccine).}

\item{perm_rise}{Numeric scalar. Permanent rise in titre following an
infection. It must have the same length as the number of infections.}

\item{temp_rise}{Numeric scalar. Initial temporary rise in titre
following an infection. It must have the same length as the number of
infections.}

\item{temp_decay}{Numeric scalar. Decay rate of the temporary titre rise.}

\item{meas_sd}{A non-negative real value that corresponds to the standard
deviation of a normal measurement model of log titres.}
}
\value{
A data frame containing simulated titre values across the specified
 sampling times.
}
\description{
Simulate antibody titre trajectories from birth
}
\examples{
simulate_titres_seropositive(
  inf_times_sbs   = c(1059.646, 2743.543, 4152.091),
  sampling_times  = c(41, 74, 122, 157, 290, 468, 787, 1192, 1553),
  age             = 10,
  subject_id      = 3,
  treatment_group = 0,
  perm_rise       = 6,
  temp_rise       = 2,
  meas_sd         = 0)
}
