% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_trajectories.R
\name{simulate_true_titre_DENV}
\alias{simulate_true_titre_DENV}
\title{Simulate true titres}
\usage{
simulate_true_titre_DENV(
  sampling_times,
  perm_rise,
  temp_rise,
  temp_decay,
  treatment_group,
  infection_times = NULL,
  vac_times = NULL
)
}
\arguments{
\item{sampling_times}{Numeric vector of sampling times (in days) relative to
the start of the study.}

\item{perm_rise}{Numeric scalar. Permanent rise in titre following an
infection. It must have the same length as the number of infections.}

\item{temp_rise}{Numeric scalar. Initial temporary rise in titre
following an infection. It must have the same length as the number of
infections.}

\item{temp_decay}{Numeric scalar. Decay rate of the temporary titre rise.}

\item{treatment_group}{Integer (0 or 1) indicating treatment group.
(0 = Placebo, 1 = Vaccine).}

\item{infection_times}{Numeric vector of times of infection events relative
to enrolment. Defaults to `NULL` if no infections.}

\item{vac_times}{Numeric vector of times of vaccination events relative to
enrolment. Defaults to `NULL` if no vaccinations.}
}
\value{
A numeric vector of length equal to \code{sampling_times},
 containing simulated log antibody titre values.
}
\description{
This function generates simulated log-transformed antibody titre values
at specified sampling times, given infection (and potentially vaccination)
events and immune response parameters. This function assumes that the
individual is seronegative at time 0. Time 0 can represent the start of a
study or the time at which the individual became susceptible.
}
\examples{
sampling_times  <- c(182, 210, 266, 294, 434, 643, 980, 1347, 1740)
simulate_true_titre(sampling_times   = sampling_times,
                    perm_rise        = 6,
                    temp_rise        = 2,
                    temp_decay       = 0.003,
                    treatment_group  = 0, # Placebo
                    infection_times  = c(300, 900))
}
