% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titre_dynamics.R
\name{titre_decay_floor}
\alias{titre_decay_floor}
\title{Linear antibody titre dynamics with lower bound}
\usage{
titre_decay_floor(par_alpha, par_beta, par_delta, time)
}
\arguments{
\item{par_alpha}{A scalar indicating the permanent rise.}

\item{par_beta}{A scalar indicating the temporary rise.}

\item{par_delta}{A scalar indicating the decay rate.}

\item{time}{Numeric vector. Time points at which to evaluate titres.}
}
\value{
Numeric vector of titres at each time point.
}
\description{
Linear antibody titre dynamics with lower bound
}
\examples{
titre_decay_floor(par_alpha = 6,
                  par_beta  = 2,
                  par_delta = 0.003,
                  time      = seq(0, 365 * 3))
}
